credentials_manager/server/requirements.txt
cryptography==41.0.3
jsonschema==4.19.0
mysql-connector-python >= 8.1.0
PyKCS11==1.5.12
bycrypt==4.0.1
setuptools==52.0.0
tabulate==0.9.0
